Move shared functions and data between sayt and products to common storybook file
